The z rotation minimization eliminated from DirectMirror if only 5 adjustable actuators are specified.
First of all, this is probably a bit of a hack. This was my way of having the fitter ignore
orientation axes that aren't allowed to vary.  Right now the minimizer will ignore the z rotation
axis if only 5 adjustable actuators have been specified. I think it is safe to assume that if there
are only 5 actuators, then there is a z rotation fixed link. However, it is foreseeable that we
might want to hold any arbitrary orientation axis constant for a minimization (like the 3.5m tert).
In any case, letting the fitter vary constrained axes is bad!!