A user specified orientaiton is adjusted. For fixed links: _physFromOrient changed. Now one common minimizer.
Orientations are now adjusted prior to determining mount coords. Orient perturbations in fixed link axes are computed first, then
included. This should make a difference in the commandable link lengths.  _physFromOrient will now return a phys value for a fixed
length link that is non-zero.  Although in reality this is impossible, we use this feature in minimization.  The same minimizer is
called when adjusting 'forward' orientations, and computing 'backward' orientations, which I like.