physFromMirPos will now return a value for Fixed Links.  Added a new test.
I changed physFromMirPos in link.py so that it can be called for a fixed link
and return a non-zero value.  This was done so I could compare the computed
mount coords using the old vs. new geometry, because the new geometry
calculation is computed in physFromMirPos, and I had bypassed this method call
previously (when it only returned 0 for fixed links) to get phys lengths from
fixed length links for minimization purposes.